begin splitting modules
