+++ tl.lua
@@ -1,509 +1,19 @@
 local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 then local p, m = pcall(require, 'compat53.module'); if p then _tl_compat = m end end; local assert = _tl_compat and _tl_compat.assert or assert; local debug = _tl_compat and _tl_compat.debug or debug; local io = _tl_compat and _tl_compat.io or io; local ipairs = _tl_compat and _tl_compat.ipairs or ipairs; local load = _tl_compat and _tl_compat.load or load; local math = _tl_compat and _tl_compat.math or math; local _tl_math_maxinteger = math.maxinteger or math.pow(2, 53); local os = _tl_compat and _tl_compat.os or os; local package = _tl_compat and _tl_compat.package or package; local pairs = _tl_compat and _tl_compat.pairs or pairs; local string = _tl_compat and _tl_compat.string or string; local table = _tl_compat and _tl_compat.table or table; local type = type; local utf8 = _tl_compat and _tl_compat.utf8 or utf8
 local VERSION = "0.24.6+dev"
 
+local errors = require("teal.errors")
 
 
 
+local lexer = require("teal.lexer")
 
 
 
-local prelude = [=====[
-do
-   global interface any
-   end
-   global interface thread
-   end
-   global interface userdata
-      is userdata
-   end
+local binary_search = require("teal.binary_search")
 
-   global record metatable<T>
-      enum Mode
-         "k" "v" "kv"
-      end
+local prelude = require("teal.embed.prelude")
+local stdlib = require("teal.embed.stdlib")
 
-      __call: function(T, any...): any...
-      __mode: Mode
-      __name: string
-      __tostring: function(T): string
-      __pairs: function<K, V>(T): function(): (K, V)
-
-      __index: any --[[FIXME: function | table | anything with an __index metamethod]]
-      __newindex: any --[[FIXME: function | table | anything with an __index metamethod]]
-
-      __gc: function(T)
-      __close: function(T)
-
-      __add: function<A, B, C>(A, B): C
-      __sub: function<A, B, C>(A, B): C
-      __mul: function<A, B, C>(A, B): C
-      __div: function<A, B, C>(A, B): C
-      __idiv: function<A, B, C>(A, B): C
-      __mod: function<A, B, C>(A, B): C
-      __pow: function<A, B, C>(A, B): C
-      __band: function<A, B, C>(A, B): C
-      __bor: function<A, B, C>(A, B): C
-      __bxor: function<A, B, C>(A, B): C
-      __shl: function<A, B, C>(A, B): C
-      __shr: function<A, B, C>(A, B): C
-      __concat: function<A, B, C>(A, B): C
-
-      __len: function<A>(T): A
-      __unm: function<A>(T): A
-      __bnot: function<A>(T): A
-
-      __eq: function<A, B>(A, B): boolean
-      __lt: function<A, B>(A, B): boolean
-      __le: function<A, B>(A, B): boolean
-   end
-end
-]=====]
-
-
-
-
-
-
-
-
-
-
-
-
-
-local stdlib = [=====[
-do
-   local enum FileStringMode
-      "a" "l" "L" "*a" "*l" "*L"
-   end
-
-   local enum FileNumberMode
-      "n" "*n"
-   end
-
-   local enum FileMode
-      "a" "l" "L" "*a" "*l" "*L" "n" "*n"
-   end
-
-   global record FILE
-      is userdata
-
-      enum SeekWhence
-         "set" "cur" "end"
-      end
-
-      enum SetVBufMode
-         "no" "full" "line"
-      end
-
-      close: function(FILE): boolean, string, integer
-      flush: function(FILE)
-
-      lines: function(FILE): (function(): (string))
-      lines: function(FILE, FileNumberMode...): (function(): (number...))
-      lines: function(FILE, (number | FileStringMode)...): (function(): (string...))
-      lines: function(FILE, (number | FileMode)...): (function(): ((string | number)...))
-      lines: function(FILE, (number | string)...): (function(): (string...))
-
-      read: function(FILE): string
-      read: function(FILE, FileNumberMode...): number...
-      read: function(FILE, (number | FileStringMode)...): string...
-      read: function(FILE, (number | FileMode)...): ((string | number)...)
-      read: function(FILE, (number | string)...): (string...)
-
-      seek: function(FILE, ? SeekWhence, ? integer): integer, string
-      setvbuf: function(FILE, SetVBufMode, ? integer)
-
-      write: function(FILE, (string | number)...): FILE, string
-
-      metamethod __close: function(FILE)
-   end
-
-   global record coroutine
-      type Function = function(any...): any...
-
-      close: function(thread): boolean, string
-      create: function(Function): thread
-      isyieldable: function(): boolean
-      resume: function(thread, any...): boolean, any...
-      running: function(): thread, boolean
-      status: function(thread): string
-      wrap: function<F>(F): F
-      yield: function(any...): any...
-   end
-
-   global record debug
-      record GetInfoTable
-         name: string
-         namewhat: string
-         source: string
-         short_src: string
-         linedefined: integer
-         lastlinedefined: integer
-         what: string
-         currentline: integer
-         istailcall: boolean
-         nups: integer
-         nparams: integer
-         isvararg: boolean
-         func: any
-         activelines: {integer:boolean}
-      end
-
-      enum HookEvent
-         "call" "tail call" "return" "line" "count"
-      end
-
-      type HookFunction = function(HookEvent, integer)
-
-      type AnyFunction = function(any...):any...
-
-      debug: function()
-      gethook: function(? thread): HookFunction, integer
-
-      getinfo: function(AnyFunction | integer): GetInfoTable
-      getinfo: function(AnyFunction | integer, string): GetInfoTable
-      getinfo: function(thread, AnyFunction | integer, string): GetInfoTable
-
-      getlocal: function(thread, AnyFunction, integer): string
-      getlocal: function(thread, integer, integer): string, any
-      getlocal: function(AnyFunction, integer): string
-      getlocal: function(integer, integer): string, any
-
-      getmetatable: function<T>(T): metatable<T>
-      getregistry: function(): {any:any}
-      getupvalue: function(AnyFunction, integer): any
-      getuservalue: function(userdata, integer): any
-
-      sethook: function(thread, HookFunction, string, ? integer)
-      sethook: function(HookFunction, string, ? integer)
-
-      setlocal: function(thread, integer, integer, any): string
-      setlocal: function(integer, integer, any): string
-
-      setmetatable: function<T>(T, metatable<T>): T
-      setupvalue: function(AnyFunction, integer, any): string
-      setuservalue: function<U>(U, any, integer): U --[[U is userdata]]
-
-      traceback: function(thread, ? string, ? integer): string
-      traceback: function(? string, ? integer): string
-
-      upvalueid: function(AnyFunction, integer): userdata
-      upvaluejoin: function(AnyFunction, integer, AnyFunction, integer)
-   end
-
-   global record io
-      enum OpenMode
-         "r" "w" "a" "r+" "w+" "a+"
-         "rb" "wb" "ab" "r+b" "w+b" "a+b"
-         "*r" "*w" "*a" "*r+" "*w+" "*a+"
-         "*rb" "*wb" "*ab" "*r+b" "*w+b" "*a+b"
-      end
-
-      close: function(? FILE)
-      input: function(? FILE | string): FILE
-      flush: function()
-
-      lines: function(? string): (function(): (string))
-      lines: function(? string, FileNumberMode...): (function(): (number...))
-      lines: function(? string, (number | FileStringMode)...): (function(): (string...))
-      lines: function(? string, (number | FileMode)...): (function(): ((string | number)...))
-      lines: function(? string, (number | string)...): (function(): (string...))
-
-      open: function(string, ? OpenMode): FILE, string, integer
-      output: function(? FILE | string): FILE
-      popen: function(string, ? OpenMode): FILE, string
-
-      read: function(): string
-      read: function(FileNumberMode...): number...
-      read: function((number | FileStringMode)...): string...
-      read: function((number | FileMode)...): ((string | number)...)
-      read: function((number | string)...): (string...)
-
-      stderr: FILE
-      stdin: FILE
-      stdout: FILE
-      tmpfile: function(): FILE
-      type: function(any): string
-      write: function((string | number)...): FILE, string
-   end
-
-   global record math
-      type Numeric = number | integer
-
-      abs: function<N is Numeric>(N): N
-      acos: function(number): number
-      asin: function(number): number
-      atan: function(number, ? number): number
-      atan2: function(number, number): number
-      ceil: function(number): integer
-      cos: function(number): number
-      cosh: function(number): number
-      deg: function(number): number
-      exp: function(number): number
-      floor: function(number): integer
-
-      fmod: function(integer, integer): integer
-      fmod: function(number, number): number
-
-      frexp: function(number): number, integer
-      huge: number
-      ldexp: function(number, integer): number
-      log: function(number, ? number): number
-      log10: function(number): number
-
-      max: function(integer...): integer
-      max: function((number | integer)...): number
-      max: function<T>(T...): T
-      max: function(any...): any
-
-      maxinteger: integer --[[needs_compat]]
-
-      min: function(integer...): integer
-      min: function((number | integer)...): number
-      min: function<T>(T...): T
-      min: function(any...): any
-
-      mininteger: integer --[[needs_compat]]
-
-      modf: function(number): integer, number
-      pi: number
-      pow: function(number, number): number
-      rad: function(number): number
-
-      random: function(integer, ? integer): integer
-      random: function(): number
-
-      randomseed: function(? integer, ? integer): integer, integer
-      sin: function(number): number
-      sinh: function(number): number
-      sqrt: function(number): number
-      tan: function(number): number
-      tanh: function(number): number
-      tointeger: function(any): integer
-      type: function(any): string
-      ult: function(number, number): boolean
-   end
-
-   global record os
-      record DateTable
-         year: integer
-         month: integer
-         day: integer
-         hour: integer
-         min: integer
-         sec: integer
-         wday: integer
-         yday: integer
-         isdst: boolean
-      end
-
-      enum DateMode
-         "!*t" "*t"
-      end
-
-      clock: function(): number
-
-      date: function(DateMode, ? number): DateTable
-      date: function(? string, ? number): string
-
-      difftime: function(integer, integer): number
-      execute: function(string): boolean, string, integer
-      exit: function(? (integer | boolean), ? boolean)
-      getenv: function(string): string
-      remove: function(string): boolean, string
-      rename: function(string, string): boolean, string
-      setlocale: function(string, ? string): string
-      time: function(? DateTable): integer
-      tmpname: function(): string
-   end
-
-   global record package
-      config: string
-      cpath: string
-      loaded: {string:any}
-      loadlib: function(string, string): (function)
-      loaders: { function(string): any, any }
-      path: string
-      preload: {any:any}
-      searchers: { function(string): any }
-      searchpath: function(string, string, ? string, ? string): string, string
-   end
-
-   global record string
-      byte: function(string, ? integer): integer
-      byte: function(string, integer, ? integer): integer...
-
-      char: function(integer...): string
-      dump: function(function(any...): (any), ? boolean): string
-      find: function(string, string, ? integer, ? boolean): integer, integer, string... --[[special_function]]
-      format: function(string, any...): string --[[special_function]]
-      gmatch: function(string, string, ? integer): (function(): string...) --[[special_function]]
-
-      gsub: function(string, string, string, ? integer): string, integer --[[special_function]]
-      gsub: function(string, string, {string:string|integer|number}, ? integer): string, integer --[[special_function]]
-      gsub: function(string, string, {integer:string|integer|number}, ? integer): string, integer --[[special_function]]
-      gsub: function(string, string, function((string|integer)...): ((string|integer|number)...), ? integer): string, integer --[[special_function]]
-
-      len: function(string): integer
-      lower: function(string): string
-      match: function(string, string, ? integer): string... --[[special_function]]
-      pack: function(string, any...): string --[[special_function]]
-      packsize: function(string): integer
-      rep: function(string, integer, ? string): string
-      reverse: function(string): string
-      sub: function(string, integer, ? integer): string
-      unpack: function(string, string, ? integer): any... --[[special_function]]
-      upper: function(string): string
-   end
-
-   global record table
-      type SortFunction = function<A>(A, A): boolean
-
-      record PackTable<A>
-         is {A}
-
-         n: integer
-      end
-
-      concat: function({(string | number)}, ? string, ? integer, ? integer): string
-
-      insert: function<A>({A}, integer, A)
-      insert: function<A>({A}, A)
-
-      move: function<A>({A}, integer, integer, integer, ? {A}): {A}
-
-      pack: function<T>(T...): PackTable<T> --[[needs_compat]]
-      pack: function(any...): {any:any} --[[needs_compat]]
-
-      remove: function<A>({A}, ? integer): A
-      sort: function<A>({A}, ? SortFunction<A>)
-
-      unpack: function<A>({A}, ? number, ? number): A... --[[needs_compat]]
-      unpack: function<A1, A2>({A1, A2}): A1, A2 --[[needs_compat]]
-      unpack: function<A1, A2, A3>({A1, A2, A3}): A1, A2, A3 --[[needs_compat]]
-      unpack: function<A1, A2, A3, A4>({A1, A2, A3, A4}): A1, A2, A3, A4 --[[needs_compat]]
-      unpack: function<A1, A2, A3, A4, A5>({A1, A2, A3, A4, A5}): A1, A2, A3, A4, A5 --[[needs_compat]]
-   end
-
-   global record utf8
-      char: function(number...): string
-      charpattern: string
-      codepoint: function(string, ? number, ? number, ? boolean): number...
-      codes: function(string, ? boolean): (function(string, ? number): (number, number))
-      len: function(string, ? number, ? number, ? boolean): number
-      offset: function(string, number, ? number): number
-   end
-
-   local record StandardLibrary
-      enum CollectGarbageCommand
-         "collect"
-         "count"
-         "stop"
-         "restart"
-      end
-
-      enum CollectGarbageSetValue
-         "step"
-         "setpause"
-         "setstepmul"
-      end
-
-      enum CollectGarbageIsRunning
-         "isrunning"
-      end
-
-      type LoadFunction = function(): string
-
-      enum LoadMode
-         "b" "t" "bt"
-      end
-
-      type XpcallMsghFunction = function(...: any): ()
-
-      arg: {string}
-      assert: function<A, B>(A, ? B, ...: any): A --[[special_function]]
-
-      collectgarbage: function(? CollectGarbageCommand): number
-      collectgarbage: function(CollectGarbageSetValue, integer): number
-      collectgarbage: function(CollectGarbageIsRunning): boolean
-      collectgarbage: function(string, ? number): (boolean | number)
-
-      dofile: function(? string): any...
-
-      error: function(? any, ? integer)
-      getmetatable: function<T>(T): metatable<T>
-      ipairs: function<A>({A}): (function():(integer, A)) --[[special_function]]
-
-      load: function((string | LoadFunction), ? string, ? LoadMode, ? table): (function, string)
-      load: function((string | LoadFunction), ? string, ? string, ? table): (function, string)
-
-      loadfile: function(? string, ? string, ? {any:any}): (function, string)
-
-      next: function<K, V>({K:V}, ? K): (K, V)
-      next: function<A>({A}, ? integer): (integer, A)
-
-      pairs: function<K, V>({K:V}): (function():(K, V)) --[[special_function]]
-      pcall: function(function(any...):(any...), any...): boolean, any... --[[special_function]]
-      print: function(any...)
-      rawequal: function(any, any): boolean
-
-      rawget: function<K, V>({K:V}, K): V --[[special_function]]
-      rawget: function({any:any}, any): any
-      rawget: function(any, any): any
-
-      rawlen: function<A>({A}): integer
-
-      rawset: function<K, V>({K:V}, K, V): {K:V}
-      rawset: function({any:any}, any, any): {any:any}
-      rawset: function(any, any, any): any
-
-      require: function(string): any --[[special_function]]
-
-      select: function<T>(integer, T...): T...
-      select: function(integer, any...): any...
-      select: function(string, any...): integer
-
-      setmetatable: function<T>(T, metatable<T>): T
-
-      tonumber: function(any): number
-      tonumber: function(any, integer): integer
-
-      tostring: function(any): string
-      type: function(any): string
-      warn: function(string, string...)
-      xpcall: function(function(any...):(any...), XpcallMsghFunction, any...): boolean, any... --[[special_function]]
-      _VERSION: string
-   end
-
-   global arg <const> = StandardLibrary.arg
-   global assert <const> = StandardLibrary.assert
-   global collectgarbage <const> = StandardLibrary.collectgarbage
-   global dofile <const> = StandardLibrary.dofile
-   global error <const> = StandardLibrary.error
-   global getmetatable <const> = StandardLibrary.getmetatable
-   global load <const> = StandardLibrary.load
-   global loadfile <const> = StandardLibrary.loadfile
-   global next <const> = StandardLibrary.next
-   global pairs <const> = StandardLibrary.pairs
-   global pcall <const> = StandardLibrary.pcall
-   global print <const> = StandardLibrary.print
-   global rawequal <const> = StandardLibrary.rawequal
-   global rawget <const> = StandardLibrary.rawget
-   global rawlen <const> = StandardLibrary.rawlen
-   global rawset <const> = StandardLibrary.rawset
-   global require <const> = StandardLibrary.require
-   global select <const> = StandardLibrary.select
-   global setmetatable <const> = StandardLibrary.setmetatable
-   global tostring <const> = StandardLibrary.tostring
-   global tonumber <const> = StandardLibrary.tonumber
-   global ipairs <const> = StandardLibrary.ipairs
-   global type <const> = StandardLibrary.type
-   global xpcall <const> = StandardLibrary.xpcall
-   global _VERSION <const> = StandardLibrary._VERSION
-end
-
-]=====]
 
 
 
@@ -518,7 +28,7 @@ local Errors = {}
 
 
 
-local tl = { GenerateOptions = {}, CheckOptions = {}, Env = {}, Result = {}, Error = {}, TypeInfo = {}, TypeReport = {}, EnvOptions = {}, Comment = {}, Token = {}, TypeCheckOptions = {} }
+local tl = { GenerateOptions = {}, CheckOptions = {}, Env = {}, Result = {}, TypeInfo = {}, TypeReport = {}, EnvOptions = {}, TypeCheckOptions = {} }
 
 
 
@@ -692,50 +202,9 @@ local tl = { GenerateOptions = {}, CheckOptions = {}, Env = {}, Result = {}, Err
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+tl.warning_kinds = errors.warning_kinds
+tl.lex = lexer.lex
+tl.get_token_at = lexer.get_token_at
 
 local TypeReporter = {}
 
@@ -745,17 +214,6 @@ local TypeReporter = {}
 
 
 
-
-local wk = {
-   ["unknown"] = true,
-   ["unused"] = true,
-   ["redeclaration"] = true,
-   ["branch"] = true,
-   ["hint"] = true,
-   ["debug"] = true,
-   ["unread"] = true,
-}
-tl.warning_kinds = wk
 
 
 
@@ -798,11 +256,6 @@ tl.typecodes = {
    UNKNOWN = 0x80008000,
    INVALID = 0x80000000,
 }
-
-
-
-
-
 
 
 
@@ -885,758 +338,6 @@ if TL_DEBUG then
             end
          end
       end, "cr", 100)
-   end
-end
-
-
-
-
-
-do
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-   local last_token_kind = {
-      ["start"] = nil,
-      ["any"] = nil,
-      ["identifier"] = "identifier",
-      ["got -"] = "op",
-      ["got --"] = nil,
-      ["got ."] = ".",
-      ["got .."] = "op",
-      ["got ="] = "op",
-      ["got ~"] = "op",
-      ["got ["] = "[",
-      ["got 0"] = "number",
-      ["got <"] = "op",
-      ["got >"] = "op",
-      ["got /"] = "op",
-      ["got :"] = "op",
-      ["got --["] = nil,
-      ["string single"] = "$ERR$",
-      ["string single got \\"] = "$ERR$",
-      ["string double"] = "$ERR$",
-      ["string double got \\"] = "$ERR$",
-      ["string long"] = "$ERR$",
-      ["string long got ]"] = "$ERR$",
-      ["comment short"] = nil,
-      ["comment long"] = "$ERR$",
-      ["comment long got ]"] = "$ERR$",
-      ["number dec"] = "integer",
-      ["number decfloat"] = "number",
-      ["number hex"] = "integer",
-      ["number hexfloat"] = "number",
-      ["number power"] = "number",
-      ["number powersign"] = "$ERR$",
-      ["pragma"] = nil,
-      ["pragma any"] = nil,
-      ["pragma word"] = "pragma_identifier",
-   }
-
-   local keywords = {
-      ["and"] = true,
-      ["break"] = true,
-      ["do"] = true,
-      ["else"] = true,
-      ["elseif"] = true,
-      ["end"] = true,
-      ["false"] = true,
-      ["for"] = true,
-      ["function"] = true,
-      ["goto"] = true,
-      ["if"] = true,
-      ["in"] = true,
-      ["local"] = true,
-      ["nil"] = true,
-      ["not"] = true,
-      ["or"] = true,
-      ["repeat"] = true,
-      ["return"] = true,
-      ["then"] = true,
-      ["true"] = true,
-      ["until"] = true,
-      ["while"] = true,
-   }
-
-   local lex_any_char_states = {
-      ["\""] = "string double",
-      ["'"] = "string single",
-      ["-"] = "got -",
-      ["."] = "got .",
-      ["0"] = "got 0",
-      ["<"] = "got <",
-      [">"] = "got >",
-      ["/"] = "got /",
-      [":"] = "got :",
-      ["="] = "got =",
-      ["~"] = "got ~",
-      ["["] = "got [",
-   }
-
-   for c = string.byte("a"), string.byte("z") do
-      lex_any_char_states[string.char(c)] = "identifier"
-   end
-   for c = string.byte("A"), string.byte("Z") do
-      lex_any_char_states[string.char(c)] = "identifier"
-   end
-   lex_any_char_states["_"] = "identifier"
-
-   for c = string.byte("1"), string.byte("9") do
-      lex_any_char_states[string.char(c)] = "number dec"
-   end
-
-   local lex_word = {}
-   for c = string.byte("a"), string.byte("z") do
-      lex_word[string.char(c)] = true
-   end
-   for c = string.byte("A"), string.byte("Z") do
-      lex_word[string.char(c)] = true
-   end
-   for c = string.byte("0"), string.byte("9") do
-      lex_word[string.char(c)] = true
-   end
-   lex_word["_"] = true
-
-   local lex_decimals = {}
-   for c = string.byte("0"), string.byte("9") do
-      lex_decimals[string.char(c)] = true
-   end
-
-   local lex_hexadecimals = {}
-   for c = string.byte("0"), string.byte("9") do
-      lex_hexadecimals[string.char(c)] = true
-   end
-   for c = string.byte("a"), string.byte("f") do
-      lex_hexadecimals[string.char(c)] = true
-   end
-   for c = string.byte("A"), string.byte("F") do
-      lex_hexadecimals[string.char(c)] = true
-   end
-
-   local lex_any_char_kinds = {}
-   local single_char_kinds = { "[", "]", "(", ")", "{", "}", ",", ";", "?" }
-   for _, c in ipairs(single_char_kinds) do
-      lex_any_char_kinds[c] = c
-   end
-   for _, c in ipairs({ "#", "+", "*", "|", "&", "%", "^" }) do
-      lex_any_char_kinds[c] = "op"
-   end
-
-   local lex_space = {}
-   for _, c in ipairs({ " ", "\t", "\v", "\n", "\r" }) do
-      lex_space[c] = true
-   end
-
-   local escapable_characters = {
-      a = true,
-      b = true,
-      f = true,
-      n = true,
-      r = true,
-      t = true,
-      v = true,
-      z = true,
-      ["\\"] = true,
-      ["\'"] = true,
-      ["\""] = true,
-      ["\r"] = true,
-      ["\n"] = true,
-   }
-
-   local function lex_string_escape(input, i, c)
-      if escapable_characters[c] then
-         return 0, true
-      elseif c == "x" then
-         return 2, (
-         lex_hexadecimals[input:sub(i + 1, i + 1)] and
-         lex_hexadecimals[input:sub(i + 2, i + 2)])
-
-      elseif c == "u" then
-         if input:sub(i + 1, i + 1) == "{" then
-            local p = i + 2
-            if not lex_hexadecimals[input:sub(p, p)] then
-               return 2, false
-            end
-            while true do
-               p = p + 1
-               c = input:sub(p, p)
-               if not lex_hexadecimals[c] then
-                  return p - i, c == "}"
-               end
-            end
-         end
-      elseif lex_decimals[c] then
-         local len = lex_decimals[input:sub(i + 1, i + 1)] and
-         (lex_decimals[input:sub(i + 2, i + 2)] and 2 or 1) or
-         0
-         return len, tonumber(input:sub(i, i + len)) < 256
-      else
-         return 0, false
-      end
-   end
-
-   tl.lex = function(input, filename)
-      local tokens = {}
-
-      local state = "any"
-      local fwd = true
-      local y = 1
-      local x = 0
-      local i = 0
-      local lc_open_lvl = 0
-      local lc_close_lvl = 0
-      local ls_open_lvl = 0
-      local ls_close_lvl = 0
-      local errs = {}
-      local nt = 0
-
-      local tx
-      local ty
-      local ti
-      local in_token = false
-
-      local comments
-
-      local function begin_token()
-         tx = x
-         ty = y
-         ti = i
-         in_token = true
-      end
-
-      local function end_token(kind, tk)
-         nt = nt + 1
-         tokens[nt] = {
-            x = tx,
-            y = ty,
-            tk = tk,
-            kind = kind,
-            comments = comments,
-         }
-         comments = nil
-         in_token = false
-      end
-
-      local function end_token_identifier()
-         local tk = input:sub(ti, i - 1)
-         nt = nt + 1
-         tokens[nt] = {
-            x = tx,
-            y = ty,
-            tk = tk,
-            kind = keywords[tk] and "keyword" or "identifier",
-            comments = comments,
-         }
-         comments = nil
-         in_token = false
-      end
-
-      local function end_token_prev(kind)
-         local tk = input:sub(ti, i - 1)
-         nt = nt + 1
-         tokens[nt] = {
-            x = tx,
-            y = ty,
-            tk = tk,
-            kind = kind,
-            comments = comments,
-         }
-         comments = nil
-         in_token = false
-      end
-
-      local function end_token_here(kind)
-         local tk = input:sub(ti, i)
-         nt = nt + 1
-         tokens[nt] = {
-            x = tx,
-            y = ty,
-            tk = tk,
-            kind = kind,
-            comments = comments,
-
-         }
-         comments = nil
-         in_token = false
-      end
-
-      local function drop_token()
-         in_token = false
-      end
-
-      local function add_syntax_error(msg)
-         local t = tokens[nt]
-         table.insert(errs, {
-            filename = filename,
-            y = t.y,
-            x = t.x,
-            msg = msg or "invalid token '" .. t.tk .. "'",
-         })
-      end
-
-      local function add_comment(text)
-         if not comments then
-            comments = {}
-         end
-         comments[#comments + 1] = {
-            x = tx,
-            y = ty,
-            text = text,
-         }
-      end
-
-      local len = #input
-      if input:sub(1, 2) == "#!" then
-         begin_token()
-         i = input:find("\n")
-         if not i then
-            i = len + 1
-         end
-         end_token_prev("hashbang")
-         y = 2
-         x = 0
-      end
-      state = "any"
-
-      while i <= len do
-         if fwd then
-            i = i + 1
-            if i > len then
-               break
-            end
-         end
-
-         local c = input:sub(i, i)
-
-         if fwd then
-            if c == "\n" then
-               y = y + 1
-               x = 0
-            else
-               x = x + 1
-            end
-         else
-            fwd = true
-         end
-
-         if state == "any" then
-            local st = lex_any_char_states[c]
-            if st then
-               state = st
-               begin_token()
-            else
-               local k = lex_any_char_kinds[c]
-               if k then
-                  begin_token()
-                  end_token(k, c)
-               elseif not lex_space[c] then
-                  begin_token()
-                  end_token_here("$ERR$")
-                  add_syntax_error()
-               end
-            end
-         elseif state == "identifier" then
-            if not lex_word[c] then
-               end_token_identifier()
-               fwd = false
-               state = "any"
-            end
-         elseif state == "string double" then
-            if c == "\\" then
-               state = "string double got \\"
-            elseif c == "\"" then
-               end_token_here("string")
-               state = "any"
-            end
-         elseif state == "comment short" then
-            if c == "\n" then
-               add_comment(input:sub(ti, i - 1))
-               state = "any"
-            end
-         elseif state == "got =" then
-            local t
-            if c == "=" then
-               t = "=="
-            else
-               t = "="
-               fwd = false
-            end
-            end_token("op", t)
-            state = "any"
-         elseif state == "got ." then
-            if c == "." then
-               state = "got .."
-            elseif lex_decimals[c] then
-               state = "number decfloat"
-            else
-               end_token(".", ".")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "got :" then
-            local t
-            if c == ":" then
-               t = "::"
-            else
-               t = ":"
-               fwd = false
-            end
-            end_token(t, t)
-            state = "any"
-         elseif state == "got [" then
-            if c == "[" then
-               state = "string long"
-            elseif c == "=" then
-               ls_open_lvl = ls_open_lvl + 1
-            else
-               end_token("[", "[")
-               fwd = false
-               state = "any"
-               ls_open_lvl = 0
-            end
-         elseif state == "number dec" then
-            if lex_decimals[c] then
-
-            elseif c == "." then
-               state = "number decfloat"
-            elseif c == "e" or c == "E" then
-               state = "number powersign"
-            else
-               end_token_prev("integer")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "got -" then
-            if c == "-" then
-               state = "got --"
-            else
-               end_token("op", "-")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "got .." then
-            if c == "." then
-               end_token("...", "...")
-            else
-               end_token("op", "..")
-               fwd = false
-            end
-            state = "any"
-         elseif state == "number hex" then
-            if lex_hexadecimals[c] then
-
-            elseif c == "." then
-               state = "number hexfloat"
-            elseif c == "p" or c == "P" then
-               state = "number powersign"
-            else
-               end_token_prev("integer")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "got --" then
-            if c == "[" then
-               state = "got --["
-            elseif c == "#" then
-               state = "pragma"
-            else
-               fwd = false
-               state = "comment short"
-               drop_token()
-            end
-         elseif state == "pragma" then
-            if not lex_word[c] then
-               end_token_prev("pragma")
-               if tokens[nt].tk == "--#pragma" then
-                  state = "pragma any"
-               else
-                  state = "comment short"
-                  table.remove(tokens)
-                  nt = nt - 1
-                  drop_token()
-               end
-               fwd = false
-            end
-         elseif state == "pragma any" then
-            if c == "\n" then
-               state = "any"
-            elseif lex_word[c] then
-               state = "pragma word"
-               begin_token()
-            elseif not lex_space[c] then
-               begin_token()
-               end_token_here("$ERR$")
-               add_syntax_error()
-            end
-         elseif state == "pragma word" then
-            if not lex_word[c] then
-               end_token_prev("pragma_identifier")
-               fwd = false
-               state = (c == "\n") and "any" or "pragma any"
-            end
-         elseif state == "got 0" then
-            if c == "x" or c == "X" then
-               state = "number hex"
-            elseif c == "e" or c == "E" then
-               state = "number powersign"
-            elseif lex_decimals[c] then
-               state = "number dec"
-            elseif c == "." then
-               state = "number decfloat"
-            else
-               end_token_prev("integer")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "got --[" then
-            if c == "[" then
-               state = "comment long"
-            elseif c == "=" then
-               lc_open_lvl = lc_open_lvl + 1
-            else
-               fwd = false
-               state = "comment short"
-               drop_token()
-               lc_open_lvl = 0
-            end
-         elseif state == "comment long" then
-            if c == "]" then
-               state = "comment long got ]"
-            end
-         elseif state == "comment long got ]" then
-            if c == "]" and lc_close_lvl == lc_open_lvl then
-               add_comment(input:sub(ti, i))
-               drop_token()
-               state = "any"
-               lc_open_lvl = 0
-               lc_close_lvl = 0
-            elseif c == "=" then
-               lc_close_lvl = lc_close_lvl + 1
-            else
-               state = "comment long"
-               lc_close_lvl = 0
-            end
-         elseif state == "string double got \\" then
-            local skip, valid = lex_string_escape(input, i, c)
-            i = i + skip
-            if not valid then
-               end_token_here("$ERR$")
-               add_syntax_error("malformed string")
-            end
-            x = x + skip
-            state = "string double"
-         elseif state == "string single" then
-            if c == "\\" then
-               state = "string single got \\"
-            elseif c == "'" then
-               end_token_here("string")
-               state = "any"
-            end
-         elseif state == "string single got \\" then
-            local skip, valid = lex_string_escape(input, i, c)
-            i = i + skip
-            if not valid then
-               end_token_here("$ERR$")
-               add_syntax_error("malformed string")
-            end
-            x = x + skip
-            state = "string single"
-         elseif state == "got ~" then
-            local t
-            if c == "=" then
-               t = "~="
-            else
-               t = "~"
-               fwd = false
-            end
-            end_token("op", t)
-            state = "any"
-         elseif state == "got <" then
-            local t
-            if c == "=" then
-               t = "<="
-            elseif c == "<" then
-               t = "<<"
-            else
-               t = "<"
-               fwd = false
-            end
-            end_token("op", t)
-            state = "any"
-         elseif state == "got >" then
-            local t
-            if c == "=" then
-               t = ">="
-            elseif c == ">" then
-               t = ">>"
-            else
-               t = ">"
-               fwd = false
-            end
-            end_token("op", t)
-            state = "any"
-         elseif state == "got /" then
-            local t
-            if c == "/" then
-               t = "//"
-            else
-               t = "/"
-               fwd = false
-            end
-            end_token("op", t)
-            state = "any"
-         elseif state == "string long" then
-            if c == "]" then
-               state = "string long got ]"
-            end
-         elseif state == "string long got ]" then
-            if c == "]" then
-               if ls_close_lvl == ls_open_lvl then
-                  end_token_here("string")
-                  state = "any"
-                  ls_open_lvl = 0
-                  ls_close_lvl = 0
-               end
-            elseif c == "=" then
-               ls_close_lvl = ls_close_lvl + 1
-            else
-               state = "string long"
-               ls_close_lvl = 0
-            end
-         elseif state == "number hexfloat" then
-            if c == "p" or c == "P" then
-               state = "number powersign"
-            elseif not lex_hexadecimals[c] then
-               end_token_prev("number")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "number decfloat" then
-            if c == "e" or c == "E" then
-               state = "number powersign"
-            elseif not lex_decimals[c] then
-               end_token_prev("number")
-               fwd = false
-               state = "any"
-            end
-         elseif state == "number powersign" then
-            if c == "-" or c == "+" then
-               state = "number power"
-            elseif lex_decimals[c] then
-               state = "number power"
-            else
-               end_token_here("$ERR$")
-               add_syntax_error("malformed number")
-               state = "any"
-            end
-         elseif state == "number power" then
-            if not lex_decimals[c] then
-               end_token_prev("number")
-               fwd = false
-               state = "any"
-            end
-         end
-      end
-
-      if in_token then
-         if last_token_kind[state] then
-            end_token_prev(last_token_kind[state])
-            if last_token_kind[state] == "$ERR$" then
-               local state_type = state:sub(1, 6)
-               if state_type == "string" then
-                  add_syntax_error("malformed string")
-               elseif state_type == "number" then
-                  add_syntax_error("malformed number")
-               elseif state_type == "commen" then
-                  add_syntax_error("unfinished long comment")
-               else
-                  add_syntax_error()
-               end
-            elseif keywords[tokens[nt].tk] then
-               tokens[nt].kind = "keyword"
-            end
-         else
-            drop_token()
-         end
-      end
-
-      table.insert(tokens, { x = x + 1, y = y, i = i, tk = "$EOF$", kind = "$EOF$", comments = comments })
-
-      return tokens, errs
-   end
-end
-
-local function binary_search(list, item, cmp)
-   local len = #list
-   local mid
-   local s, e = 1, len
-   while s <= e do
-      mid = math.floor((s + e) / 2)
-      local val = list[mid]
-      local res = cmp(val, item)
-      if res then
-         if mid == len then
-            return mid, val
-         else
-            if not cmp(list[mid + 1], item) then
-               return mid, val
-            end
-         end
-         s = mid + 1
-      else
-         e = mid - 1
-      end
-   end
-end
-
-tl.get_token_at = function(tks, y, x)
-   local _, found = binary_search(
-   tks, nil,
-   function(tk)
-      return tk.y < y or
-      (tk.y == y and tk.x <= x)
-   end)
-
-
-   if found and
-      found.y == y and
-      found.x <= x and x < found.x + #found.tk then
-
-      return found.tk
    end
 end
 
@@ -2387,13 +1088,13 @@ local function shallow_copy_table(t)
 end
 
 
-local function clear_redundant_errors(errors)
+local function clear_redundant_errors(errs)
    local redundant = {}
    local lastx, lasty = 0, 0
-   for i, err in ipairs(errors) do
+   for i, err in ipairs(errs) do
       err.i = i
    end
-   table.sort(errors, function(a, b)
+   table.sort(errs, function(a, b)
       local af = assert(a.filename)
       local bf = assert(b.filename)
       return af < bf or
@@ -2401,7 +1102,7 @@ local function clear_redundant_errors(errors)
       (a.y == b.y and (a.x < b.x or
       (a.x == b.x and (a.i < b.i))))))
    end)
-   for i, err in ipairs(errors) do
+   for i, err in ipairs(errs) do
       err.i = nil
       if err.x == lastx and err.y == lasty then
          table.insert(redundant, i)
@@ -2409,7 +1110,7 @@ local function clear_redundant_errors(errors)
       lastx, lasty = err.x, err.y
    end
    for i = #redundant, 1, -1 do
-      table.remove(errors, redundant[i])
+      table.remove(errs, redundant[i])
    end
 end
 
@@ -4632,7 +3333,7 @@ do
    end
 
    function tl.parse(input, filename, parse_lang)
-      local tokens, errs = tl.lex(input, filename)
+      local tokens, errs = lexer.lex(input, filename)
       local node, required_modules = tl.parse_program(tokens, errs, filename, parse_lang)
       return node, errs, required_modules
    end
@@ -7446,10 +6147,10 @@ local stdlib_globals = nil
 local globals_typeid = new_typeid()
 local fresh_typevar_ctr = 1
 
-local function assert_no_errors(errors, msg)
-   if #errors ~= 0 then
+local function assert_no_errors(errs, msg)
+   if #errs ~= 0 then
       local out = {}
-      for _, err in ipairs(errors) do
+      for _, err in ipairs(errs) do
          table.insert(out, err.y .. ":" .. err.x .. " " .. err.msg .. "\n")
       end
       error("Internal Compiler Error: " .. msg .. ":\n" .. table.concat(out), 2)
